+++ src/features/cabins/useEditCabin.js
@@ -1,0 +1,21 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { createOrEditCabin } from "../../services/apiCabins";
+import { toast } from "react-hot-toast";
+
+export function useEditCabin() {
+  const queryClient = useQueryClient();
+
+  // edit cabin : 需要參數
+  const { isLoading: isEditing, mutate: editCabin } = useMutation({
+    mutationFn: ({ newCabinData, id }) => createOrEditCabin(newCabinData, id),
+    onSuccess: () => {
+      toast.success("Cabin successfully edited!");
+      queryClient.invalidateQueries({
+        queryKey: ["cabins"],
+      });
+    },
+    onError: (err) => toast.error(err.message),
+  });
+
+  return { isEditing, editCabin}
+}
